+++ lib/ruby-next/language/rewriters/abstract.rb
@@ -4,6 +4,10 @@ module RubyNext
   module Language
     module Rewriters
       class Abstract < ::Parser::TreeRewriter
+        NAME = "custom-rewriter"
+        SYNTAX_PROBE = "1"
+        MIN_SUPPORTED_VERSION = Gem::Version.new("2.2.0")
+
         class << self
           # Returns true if the syntax is not supported
           # by the current Ruby (performs syntax check, not version check)